+++ src/icalendar/tools.py
@@ -1,29 +1,62 @@
+"""Utility functions for icalendar."""
+
 from __future__ import annotations
+
 import random
 from datetime import date, datetime, tzinfo
 from string import ascii_letters, digits
+from warnings import warn
 
 from icalendar.parser_tools import to_unicode
 
+from .error import WillBeRemovedInVersion7
+
 
 class UIDGenerator:
-    """If you are too lazy to create real uid's."""
+    """Use this only if you're too lazy to create real UUIDs.
+    
+    .. deprecated:: 6.2.1
+
+        Use the Python standard library's :func:`uuid.uuid4` instead.
+    """
 
     chars = list(ascii_letters + digits)
 
     @staticmethod
-    def rnd_string(length=16):
-        """Generates a string with random characters of length."""
+    def rnd_string(length=16) -> str:
+        """Generates a string with random characters of length.
+
+         .. deprecated:: 6.2.1
+
+             Use the Python standard library's :func:`uuid.uuid4` instead.
+        """
+        warn(
+            "Use https://docs.python.org/3/library/uuid.html#uuid.uuid4 instead.",
+            WillBeRemovedInVersion7,
+            stacklevel=1
+        )
         return "".join([random.choice(UIDGenerator.chars) for _ in range(length)])
 
     @staticmethod
     def uid(host_name="example.com", unique=""):
-        """Generates a unique id consisting of:
-            datetime-uniquevalue@host.
-        Like:
-            20050105T225746Z-HKtJMqUgdO0jDUwm@example.com
+        """Generates a unique ID consisting of ``datetime-uniquevalue@host``.
+
+        For example:
+            
+            .. code-block:: text
+
+                20050105T225746Z-HKtJMqUgdO0jDUwm@example.com
+        
+        .. deprecated:: 6.2.1
+
+            Use the Python standard library's :func:`uuid.uuid5` instead.
         """
         from icalendar.prop import vDatetime, vText
+        warn(
+            "Use https://docs.python.org/3/library/uuid.html#uuid.uuid5 instead.",
+            WillBeRemovedInVersion7,
+            stacklevel=1
+        )
 
         host_name = to_unicode(host_name)
         unique = unique or UIDGenerator.rnd_string()