+++ CHANGES.rst
@@ -20,7 +20,7 @@ Breaking changes:
 
 New features:
 
-- ...
+- Add the ``uid`` property to ``Alarm``, ``Event``, ``Calendar``, ``Todo``, and ``Journal`` components. See `Issue 740 <https://github.com/collective/icalendar/issues/740>`_.
 
 Bug fixes: