+++ CHANGES.rst
@@ -26,7 +26,7 @@ Bug fixes:
 
 - Test that we can add an RRULE as a string. See `Issue 301 <https://github.com/collective/icalendar/issues/301>`_.
 - Test that we support dateutil timezones as outlined in `Issue 336 <https://github.com/collective/icalendar/issues/336>`_.
-- Build documentation on readthedocs with the version identifier. See `Issue 826 <https://github.com/collective/icalendar/issues/826>`_.
+- Build documentation on Read the Docs with the version identifier. See `Issue 826 <https://github.com/collective/icalendar/issues/826>`_.
 
 
 6.2.0 (2025-05-07)